+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JVY8D1XBKAGSAXZ0ADKPGPA4"
+        "01JVYENB9V1Q7WHESH11WX8KQK"
     )
     r"""The id of the resource"""